OctagonalRR_AOPS_Fig_04bc.py
The code for the research presented in the paper titled "Inverse design of plasmon-based nonlinear octagonal resonators enabled by deep neural networks."
This code corresponds to the article's Forward Deep Neural Network (DNN) section.
This code regenerates the Fig. 4b and c of the paper.
Please cite the paper in any publication using this code.