OctagonalRR_AOPS_Fig_S04.py
The code for the research presented in the paper titled "Inverse design of plasmon-based nonlinear octagonal resonators enabled by deep neural networks."

This code corresponds to Section S2 (Dataset discussion) of the Supplementary Information of the article.
This code regenerates Fig. S4a and b of the paper's Supplementary Information.
Please cite the paper in any publication using this code.