OctagonalRR_AOPS_Fig_S07_drop.m
The code for the research presented in the paper titled "Inverse design of plasmon-based nonlinear octagonal resonators enabled by deep neural networks."

In this code, first and second-order derivatives of transmission spectrum versus wavelength will be calculated.
In this code, local minima and maxima will be calculated.
In this code, we proposed an algorithm to find an optimal point for switching applications.
This code regenerates the Fig. S7 of the paper's Supplementary Information.
Please cite the paper in any publication using this code.