OctagonalRR_AOPS_Fig_07bc.py
The code for the research presented in the paper titled "Inverse design of plasmon-based nonlinear octagonal resonators enabled by deep neural networks."

This code corresponds to the article's inverse Deep Neural Network (DNN) section.
This code regenerates the Fig. 7b and c of the paper.
Please cite the paper in any publication using this code.